gs_usb: make dummy endpoint selecatble by KConfig
This change adds dummy endpoint KConfig select flag
to support dynamic endpoint enumeration made in
newer version of gs_usb driver in linux kernel.

Additional effect of this change is no need of
dummy usb endpoint at all if newer version of gs_usb
driver is used.
